aaaaaaaa working on loading and saving buuuuuuuuuuut im probably gonna have to rewrite a bunch of stuff bc pain is here

+++ src/com/marcel/Rendering/Utils/DPos.java
@@ -9,7 +9,7 @@ public class DPos {
         y = 0;
     }
 
-    public DPos(float x, float y)
+    public DPos(double x, double y)
     {
         this.x = x;
         this.y = y;
